# docs/swagger.yaml
basePath: /
definitions:
  apistruct.ApiResponse:
    properties:
      data: {}
      errCode:
        example: 0
        type: integer
      errDlt:
        example: ""
        type: string
      errMsg:
        example: success
        type: string
    type: object
  apistruct.ForceLogoutReq:
    properties:
      platformID:
        example: 1
        maximum: 12
        minimum: 1
        type: integer
      userID:
        example: user123
        type: string
    required:
    - platformID
    - userID
    type: object
  apistruct.GetAdminTokenReq:
    properties:
      adminID:
        example: admin
        type: string
      secret:
        example: admin_secret
        type: string
    required:
    - adminID
    - secret
    type: object
  apistruct.GetAdminTokenResp:
    properties:
      expireTimeSeconds:
        example: 86400
        type: integer
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
...
  apistruct.GetUserTokenReq:
    properties:
      platformID:
        example: 1
        maximum: 12
        minimum: 1
        type: integer
      secret:
        example: openIM123
        type: string
      userID:
        example: user123
        type: string
    required:
    - platformID
    - secret
    - userID
    type: object
  apistruct.GetUserTokenResp:
    properties:
      expireTimeSeconds:
        example: 86400
        type: integer
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
...
  apistruct.ParseTokenReq:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    required:
    - token
    type: object
  apistruct.ParseTokenResp:
    properties:
      expireTimeSeconds:
        example: 86400
        type: integer
      platform:
        example: iOS
        type: string
      userID:
        example: user123
        type: string
    type: object
...
host: localhost:10002
info:
  contact:
    email: contact@openim.io
    name: OpenIM Support
    url: https://github.com/openimsdk/open-im-server
  description: |-
    OpenIM Server 提供即时通讯服务的 REST API 接口文档

    ## 认证方式
    大部分接口需要在 Header 中携带 token 进行认证：
    - Header名称: token
    - 获取方式: 调用 /auth/get_user_token 接口

    ## 接口规范
    - 所有接口均使用 POST 方法
    - Content-Type: application/json
    - 响应格式: JSON
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: OpenIM Server API
  version: "3.8"
paths:
  /auth/force_logout:
    post:
      consumes:
      - application/json
      description: 强制指定用户登出，清除其所有Token
      parameters:
      - description: 强制登出信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apistruct.ForceLogoutReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
        "401":
          description: 权限不足
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
      security:
      - BearerToken: []
      summary: 强制登出用户
      tags:
      - 认证管理
  /auth/get_admin_token:
    post:
      consumes:
      - application/json
      description: 管理员登录获取访问令牌
      parameters:
      - description: 管理员登录信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apistruct.GetAdminTokenReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            allOf:
            - $ref: '#/definitions/apistruct.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/apistruct.GetAdminTokenResp'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
        "401":
          description: 认证失败
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
      summary: 获取管理员Token
      tags:
      - 认证管理
  /auth/get_user_token:
    post:
      consumes:
      - application/json
      description: 用户登录获取访问令牌，用于后续API调用的身份认证
      parameters:
      - description: 用户登录信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apistruct.GetUserTokenReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应，包含token和过期时间
          schema:
            allOf:
            - $ref: '#/definitions/apistruct.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/apistruct.GetUserTokenResp'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
        "401":
          description: 认证失败，用户名或密码错误
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
      summary: 获取用户Token
      tags:
      - 认证管理
  /auth/parse_token:
    post:
      consumes:
      - application/json
      description: 解析并验证Token的有效性，返回Token中包含的用户信息
      parameters:
      - description: Token信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apistruct.ParseTokenReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应，包含用户ID和过期时间
          schema:
            allOf:
            - $ref: '#/definitions/apistruct.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/apistruct.ParseTokenResp'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
        "401":
          description: Token无效或已过期
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
      summary: 解析Token
      tags:
      - 认证管理
...
schemes:
- http
- https
securityDefinitions:
  BearerToken:
    in: header
    name: token
    type: apiKey
swagger: "2.0"